Домашнее задание 7 урок "Гид по турам" багфикс

+++ src/lesson7/CheapTripDB.java
@@ -28,7 +28,7 @@ public class CheapTripDB {
         }
     }
 
-    public void searchByPrice(double price) {
+    public void searchByPrice(int price) {
         for (Tour tour : tours) {
             if (tour.getPrice() <= price) {
                 System.out.println(tour);
@@ -36,7 +36,7 @@ public class CheapTripDB {
         }
     }
 
-    public void searchByPriceAndStars(double price, int stars) {
+    public void searchByPriceAndStars(int price, int stars) {
         for (Tour tour : tours) {
             if (tour.getPrice() <= price && tour.getStars() == stars) {
                 System.out.println(tour);
